+++ DocClass/CONSTANS.py
@@ -1,0 +1,41 @@
+"""
+Constants for documets expressions
+"""
+
+__name__ = "constants"
+__author__ = "Ehud Barda"
+
+import json
+import logging
+import os
+
+# import utils.log
+procedure_types_EXP = {
+    "Pathology and Surgery" : ["Pathology", "CYTOPATHOLOGY", "Surgical Procedure", "SURGICAL", "Surgical","PATHOLOGY"],
+    "Medication prescriptions": ["Medication", "Medications"],
+    "Imaging": ["CT", "XR", "NM", "FDG", "PET/CT", "PET", 'XR', 'MRI', 'PET CT', 'CT','DOTATATE']
+
+}
+patient_name = [],
+
+date_of_document = [],
+
+date_of_procedure = [],
+
+doctor_name = [],
+
+department= [],
+
+institution = [],
+
+procedure_type = {
+
+                 },
+
+num_of_pages = 0,
+
+document_reference = ""
+
+
+
+
